storageconfig: move StoreEncryptionSpec to cli
The separate encryption specs are a byproduct of having to use a
separate command-line flag. This change moves the related code to the
cli package.

Epic: none
Release note: None
