+++ pkg/base/store_spec.go
@@ -333,64 +333,3 @@ func (ssl *StoreSpecList) Set(value string) error {
 	}
 	return nil
 }
-
-// PopulateWithEncryptionOpts iterates through the EncryptionSpecList and looks
-// for matching paths in the StoreSpecList and WAL failover config. Any
-// unmatched EncryptionSpec causes an error.
-func PopulateWithEncryptionOpts(
-	storeSpecs StoreSpecList,
-	walFailoverConfig *storageconfig.WALFailover,
-	encryptionSpecs storageconfig.EncryptionSpecList,
-) error {
-	for _, es := range encryptionSpecs.Specs {
-		var storeMatched bool
-		for i := range storeSpecs.Specs {
-			if !es.PathMatches(storeSpecs.Specs[i].Path) {
-				continue
-			}
-
-			// Found a matching path.
-			if storeSpecs.Specs[i].EncryptionOptions != nil {
-				return fmt.Errorf("store with path %s already has an encryption setting",
-					storeSpecs.Specs[i].Path)
-			}
-
-			storeSpecs.Specs[i].EncryptionOptions = &es.Options
-			storeMatched = true
-			break
-		}
-		pathMatched, err := maybeSetExternalPathEncryption(&walFailoverConfig.Path, es)
-		if err != nil {
-			return err
-		}
-		prevPathMatched, err := maybeSetExternalPathEncryption(&walFailoverConfig.PrevPath, es)
-		if err != nil {
-			return err
-		}
-		if !storeMatched && !pathMatched && !prevPathMatched {
-			return fmt.Errorf("no usage of path %s found for encryption setting: %v", es.Path, es)
-		}
-	}
-	return nil
-}
-
-// maybeSetExternalPathEncryption updates an ExternalPath to contain the provided
-// encryption options if the path matches. The ExternalPath must not already have
-// an encryption setting.
-func maybeSetExternalPathEncryption(
-	externalPath *storageconfig.ExternalPath, es storageconfig.StoreEncryptionSpec,
-) (found bool, err error) {
-	if !externalPath.IsSet() || !es.PathMatches(externalPath.Path) {
-		return false, nil
-	}
-	// NB: The external paths WALFailoverConfig.Path and
-	// WALFailoverConfig.PrevPath are only ever set in single-store
-	// configurations. In multi-store with among-stores failover mode, these
-	// will be empty (so we won't encounter the same path twice).
-	if externalPath.Encryption != nil {
-		return false, fmt.Errorf("WAL failover path %s already has an encryption setting",
-			externalPath.Path)
-	}
-	externalPath.Encryption = &es.Options
-	return true, nil
-}
