+++ pkg/cli/flags_util.go
@@ -6,17 +6,23 @@
 package cli
 
 import (
+	"bytes"
 	gohex "encoding/hex"
 	"fmt"
 	"math"
+	"path/filepath"
 	"regexp"
 	"strconv"
 	"strings"
+	"time"
 
+	"github.com/cockroachdb/cockroach/pkg/base"
+	"github.com/cockroachdb/cockroach/pkg/cli/cliflags"
 	"github.com/cockroachdb/cockroach/pkg/keys"
 	"github.com/cockroachdb/cockroach/pkg/roachpb"
 	"github.com/cockroachdb/cockroach/pkg/server/status"
 	"github.com/cockroachdb/cockroach/pkg/storage"
+	"github.com/cockroachdb/cockroach/pkg/storage/storageconfig"
 	"github.com/cockroachdb/cockroach/pkg/util"
 	"github.com/cockroachdb/cockroach/pkg/util/humanizeutil"
 	"github.com/cockroachdb/cockroach/pkg/util/keysutil"
@@ -441,4 +447,262 @@ func (b *bytesOrPercentageValue) SafeFormat(p redact.SafePrinter, _ rune) {
 // IsSet returns true iff Set has successfully been called.
 func (b *bytesOrPercentageValue) IsSet() bool {
 	return b.bval.IsSet()
+}
+
+// populateWithEncryptionOpts iterates through the encryptionSpecList and looks
+// for matching paths in the StoreSpecList and WAL failover config. Any
+// unmatched EncryptionSpec causes an error.
+func populateWithEncryptionOpts(
+	storeSpecs base.StoreSpecList,
+	walFailoverConfig *storageconfig.WALFailover,
+	encryptionSpecs encryptionSpecList,
+) error {
+	for _, es := range encryptionSpecs.Specs {
+		var storeMatched bool
+		for i := range storeSpecs.Specs {
+			if !es.PathMatches(storeSpecs.Specs[i].Path) {
+				continue
+			}
+
+			// Found a matching path.
+			if storeSpecs.Specs[i].EncryptionOptions != nil {
+				return fmt.Errorf("store with path %s already has an encryption setting",
+					storeSpecs.Specs[i].Path)
+			}
+
+			storeSpecs.Specs[i].EncryptionOptions = &es.Options
+			storeMatched = true
+			break
+		}
+		pathMatched, err := maybeSetExternalPathEncryption(&walFailoverConfig.Path, es)
+		if err != nil {
+			return err
+		}
+		prevPathMatched, err := maybeSetExternalPathEncryption(&walFailoverConfig.PrevPath, es)
+		if err != nil {
+			return err
+		}
+		if !storeMatched && !pathMatched && !prevPathMatched {
+			return fmt.Errorf("no usage of path %s found for encryption setting: %v", es.Path, es)
+		}
+	}
+	return nil
+}
+
+// maybeSetExternalPathEncryption updates an ExternalPath to contain the provided
+// encryption options if the path matches. The ExternalPath must not already have
+// an encryption setting.
+func maybeSetExternalPathEncryption(
+	externalPath *storageconfig.ExternalPath, es storeEncryptionSpec,
+) (found bool, err error) {
+	if !externalPath.IsSet() || !es.PathMatches(externalPath.Path) {
+		return false, nil
+	}
+	// NB: The external paths WALFailoverConfig.Path and
+	// WALFailoverConfig.PrevPath are only ever set in single-store
+	// configurations. In multi-store with among-stores failover mode, these
+	// will be empty (so we won't encounter the same path twice).
+	if externalPath.Encryption != nil {
+		return false, fmt.Errorf("WAL failover path %s already has an encryption setting",
+			externalPath.Path)
+	}
+	externalPath.Encryption = &es.Options
+	return true, nil
+}
+
+// storeEncryptionSpec contains the details that can be specified in the cli via
+// the --enterprise-encryption flag.
+type storeEncryptionSpec struct {
+	Options storageconfig.EncryptionOptions
+	Path    string
+}
+
+// String returns a fully parsable version of the encryption spec.
+func (es storeEncryptionSpec) String() string {
+	// All fields are set.
+	return fmt.Sprintf("path=%s,key=%s,old-key=%s,rotation-period=%s",
+		es.Path, es.Options.KeyFiles.CurrentKey, es.Options.KeyFiles.OldKey, es.RotationPeriod(),
+	)
+}
+
+// RotationPeriod returns the rotation period as a duration.
+func (es storeEncryptionSpec) RotationPeriod() time.Duration {
+	return time.Duration(es.Options.DataKeyRotationPeriod) * time.Second
+}
+
+// PathMatches returns true if this storeEncryptionSpec matches the given store path.
+func (es storeEncryptionSpec) PathMatches(path string) bool {
+	return es.Path == path || es.Path == "*"
+}
+
+// Special value of key paths to mean "no encryption". We do not accept empty fields.
+const plaintextFieldValue = "plain"
+
+// parseStoreEncryptionSpec parses the string passed in and returns a new
+// storeEncryptionSpec if parsing succeeds.
+// TODO(mberhault): we should share the parsing code with the StoreSpec.
+func parseStoreEncryptionSpec(value string) (storeEncryptionSpec, error) {
+	const pathField = "path"
+	es := storeEncryptionSpec{
+		Path: "",
+		Options: storageconfig.EncryptionOptions{
+			KeySource:             storageconfig.EncryptionKeyFromFiles,
+			KeyFiles:              &storageconfig.EncryptionKeyFiles{},
+			DataKeyRotationPeriod: int64(storageconfig.DefaultRotationPeriod / time.Second),
+		},
+	}
+
+	used := make(map[string]struct{})
+	for _, split := range strings.Split(value, ",") {
+		if len(split) == 0 {
+			continue
+		}
+		subSplits := strings.SplitN(split, "=", 2)
+		if len(subSplits) == 1 {
+			return storeEncryptionSpec{}, fmt.Errorf("field not in the form <key>=<value>: %s", split)
+		}
+		field := strings.ToLower(subSplits[0])
+		value := subSplits[1]
+		if _, ok := used[field]; ok {
+			return storeEncryptionSpec{}, fmt.Errorf("%s field was used twice in encryption definition", field)
+		}
+		used[field] = struct{}{}
+
+		if len(field) == 0 {
+			return storeEncryptionSpec{}, fmt.Errorf("empty field")
+		}
+		if len(value) == 0 {
+			return storeEncryptionSpec{}, fmt.Errorf("no value specified for %s", field)
+		}
+
+		switch field {
+		case pathField:
+			if value == "*" {
+				es.Path = value
+			} else {
+				var err error
+				es.Path, err = getAbsoluteFSPath(pathField, value)
+				if err != nil {
+					return storeEncryptionSpec{}, err
+				}
+			}
+		case "key":
+			if value == plaintextFieldValue {
+				es.Options.KeyFiles.CurrentKey = plaintextFieldValue
+			} else {
+				var err error
+				es.Options.KeyFiles.CurrentKey, err = getAbsoluteFSPath("key", value)
+				if err != nil {
+					return storeEncryptionSpec{}, err
+				}
+			}
+		case "old-key":
+			if value == plaintextFieldValue {
+				es.Options.KeyFiles.OldKey = plaintextFieldValue
+			} else {
+				var err error
+				es.Options.KeyFiles.OldKey, err = getAbsoluteFSPath("old-key", value)
+				if err != nil {
+					return storeEncryptionSpec{}, err
+				}
+			}
+		case "rotation-period":
+			dur, err := time.ParseDuration(value)
+			if err != nil {
+				return storeEncryptionSpec{}, errors.Wrapf(err, "could not parse rotation-duration value: %s", value)
+			}
+			es.Options.DataKeyRotationPeriod = int64(dur / time.Second)
+		default:
+			return storeEncryptionSpec{}, fmt.Errorf("%s is not a valid enterprise-encryption field", field)
+		}
+	}
+
+	// Check that all fields are set.
+	if es.Path == "" {
+		return storeEncryptionSpec{}, fmt.Errorf("no path specified")
+	}
+	if es.Options.KeyFiles.CurrentKey == "" {
+		return storeEncryptionSpec{}, fmt.Errorf("no key specified")
+	}
+	if es.Options.KeyFiles.OldKey == "" {
+		return storeEncryptionSpec{}, fmt.Errorf("no old-key specified")
+	}
+
+	return es, nil
+}
+
+// encryptionSpecList contains a slice of StoreEncryptionSpecs that implements pflag's value
+// interface.
+type encryptionSpecList struct {
+	Specs []storeEncryptionSpec
+}
+
+var _ pflag.Value = &encryptionSpecList{}
+
+// String returns a string representation of all the StoreEncryptionSpecs. This is part
+// of pflag's value interface.
+func (encl encryptionSpecList) String() string {
+	var buffer bytes.Buffer
+	for _, ss := range encl.Specs {
+		fmt.Fprintf(&buffer, "--%s=%s ", cliflags.EnterpriseEncryption.Name, ss)
+	}
+	// Trim the extra space from the end if it exists.
+	if l := buffer.Len(); l > 0 {
+		buffer.Truncate(l - 1)
+	}
+	return buffer.String()
+}
+
+// Type returns the underlying type in string form. This is part of pflag's
+// value interface.
+func (encl *encryptionSpecList) Type() string {
+	return "EncryptionSpec"
+}
+
+// Set adds a new value to the StoreEncryptionSpecValue. It is the important part of
+// pflag's value interface.
+func (encl *encryptionSpecList) Set(value string) error {
+	spec, err := parseStoreEncryptionSpec(value)
+	if err != nil {
+		return err
+	}
+	if encl.Specs == nil {
+		encl.Specs = []storeEncryptionSpec{spec}
+	} else {
+		encl.Specs = append(encl.Specs, spec)
+	}
+	return nil
+}
+
+// encryptionOptionsForStore takes a store directory and returns its EncryptionOptions
+// if a matching entry if found in the StoreEncryptionSpecList.
+func encryptionOptionsForStore(
+	dir string, encryptionSpecs encryptionSpecList,
+) (*storageconfig.EncryptionOptions, error) {
+	// We need an absolute path, but the input may have come in relative.
+	path, err := filepath.Abs(dir)
+	if err != nil {
+		return nil, errors.Wrapf(err, "could not find absolute path for %s ", dir)
+	}
+	for _, es := range encryptionSpecs.Specs {
+		if es.PathMatches(path) {
+			return &es.Options, nil
+		}
+	}
+	return nil, nil
+}
+
+// getAbsoluteFSPath takes a (possibly relative) and returns the absolute path.
+// Returns an error if the path begins with '~' or Abs fails.
+// 'fieldName' is used in error strings.
+func getAbsoluteFSPath(fieldName string, p string) (string, error) {
+	if p[0] == '~' {
+		return "", fmt.Errorf("%s cannot start with '~': %s", fieldName, p)
+	}
+
+	ret, err := filepath.Abs(p)
+	if err != nil {
+		return "", errors.Wrapf(err, "could not find absolute path for %s %s", fieldName, p)
+	}
+	return ret, nil
 }
